changed to two different settings
