+++ src/model/ImportSettings.java
@@ -1,0 +1,11 @@
+package model;
+
+import java.util.List;
+
+public class ImportSettings extends Settings {
+	
+	public ImportSettings(List<String> databases) {
+		this.databases = databases;
+	}
+	
+}
